+++ package/src/cookie_consent/index.ts
@@ -3,5 +3,6 @@ export { default as useCookieConsent } from './client/use_cookie_consent';
 export { default as CookieConsentDialog } from './client/components/cookie_consent_dialog';
 export { default as PrivacyPolicyUpdateDialog } from './client/components/privacy_policy_update_dialog';
 export { default as CookieConsentAnalytics } from './client/components/cookie_consent_analytics';
+export { defaultGdprCountries } from './gdpr_countries';
 export type { CookieConsentContextType, ConsentValue, CookieDialogClassNames, CookieDialogStyles } from './types';
 export type { CookieConsentRoutingConfig, CookieConsentAnalyticsSecrets } from '../types/types';